+++ ocp/Dockerfile.buildconfig
@@ -6,4 +6,4 @@ COPY --from=openshift-origin /usr/bin/oc /usr/bin
 
 WORKDIR /srv/
 USER builder
-ENTRYPOINT ["/usr/bin/gangplank", "builder"]
+ENTRYPOINT ["/usr/bin/dumb-init", "/usr/bin/gangway"]
